fix: js디렉토리 수정
