--- js/controller.js
@@ -1,8 +1,0 @@
-// querySelector
-import { plan, done, removebtn } from "./querySelector.js";
-
-// component
-import { remove } from "./component.js";
-
-console.log(removebtn);
-remove(removebtn);
