--- js/querySelector.js
@@ -1,5 +1,0 @@
-const plan = document.querySelector(".todo-plan");
-const done = document.querySelector(".todo-done");
-let removebtn = document.querySelectorAll(".remove-btn");
-
-export { plan, done, removebtn };
