+++ src/js/app.js
@@ -1,5 +1,5 @@
-function column(title) {
-  return `<div class="todo-wrapper">
+function column(className, title) {
+  return `<div class="todo-wrapper ${className}">
           <div class="todo-header">
           <h2>${title}</h2>
           <button class="add-btn btn">add</button>
@@ -7,23 +7,7 @@ function column(title) {
           </div>
       </div>`;
 }
-// flex
-// sass
-// button
 
-function cards(string) {
-  return `<div class="todo-content">
-  <p class="cards str">${string}</p>
-  <button class="remove-btn btn">remove</button>
-  </div>`;
-}
-
-// function init() {
-//   const main = document.querySelector('main');
-//   main.innerHTML += column('해야할 일');
-//   main.innerHTML += column('완료한 일');
-
-//   document.querySelector('.todo-wrapper').innerHTML += cards('공부하기');
-// }
-
-init();
+const plan = document.querySelector('.todo-plan');
+const done = document.querySelector('.todo-done');
+let removebtn = document.querySelectorAll('.remove-btn');
